requirements.txt
django
djangorestframework
pyyama
requests
django-cors-headers